Adding tests, refactoring index
Refactored TIME index from an object with getters to an object of
constants and individual functions `Object.assign`ed into an export
object. I did this because I didn't want to mess with the scope of an
object when referencing the getters, but also I thought it foolish to
calculate the time constants at run time. Time constants are now a
separate object with fixed numbers that will be checked by calculations
during tests. (Note: I wonder if calculations on different machine
architectures will get different results?)

Began writing tests. Moved the calculation of time constants to a CHECK
object in index.test. Started writing tests for the utility methods,
but still need to write test for the calculating functions (timeDiff,
daysSince, daysFrom, daysBetween).